+++ src/main/java/org/folio/qm/service/FieldProtectionSetterService.java
@@ -4,5 +4,5 @@ import org.folio.qm.domain.dto.QuickMarc;
 
 public interface FieldProtectionSetterService {
 
-  QuickMarc applyFieldProtection(QuickMarc qmRecord);
+  void applyFieldProtection(QuickMarc qmRecord);
 }