+++ src/main/java/org/folio/qm/service/LinksService.java
@@ -1,0 +1,8 @@
+package org.folio.qm.service;
+
+import org.folio.qm.domain.dto.QuickMarc;
+
+public interface LinksService {
+
+  void setRecordLinks(QuickMarc qmRecord);
+}